7.3.5 Добавление компонента хедера
