+++ src/components/header/header.tsx
@@ -1,0 +1,69 @@
+import { Link } from 'react-router-dom';
+import Logo from '../logo/logo';
+import { AppRoute, AuthorizationStatus } from '../../const';
+import { store } from '../../store';
+import { useAppDispatch, useAppSelector } from '../../hooks';
+import { logoutAction } from '../../store/api-actions';
+
+function Header(): JSX.Element {
+  const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
+  const dispatch = useAppDispatch();
+
+  const favoriteOffers = store.getState().offers.filter((offer) => offer.isFavorite === true);
+  const userData = store.getState().user;
+
+  return (
+    <header className="header">
+      <div className="container">
+        <div className="header__wrapper">
+          <div className="header__left">
+            <Logo />
+          </div>
+          <nav className="header__nav">
+            <ul className="header__nav-list">
+              {
+                authorizationStatus === AuthorizationStatus.Auth
+                  ? (
+                    <>
+                      <li className="header__nav-item user">
+                        <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favorites}>
+                          <div className="header__avatar-wrapper user__avatar-wrapper">
+                          </div>
+                          <span className="header__user-name user__name">{userData?.email}</span>
+                          <span className="header__favorite-count">{favoriteOffers.length}</span>
+                        </Link>
+                      </li>
+                      <li className="header__nav-item">
+                        <Link
+                          className="header__nav-link"
+                          to={AppRoute.Login}
+                          onClick={(evt) => {
+                            evt.preventDefault();
+                            dispatch(logoutAction());
+                          }}
+                        >
+                          <span className="header__signout">Sign out</span>
+                        </Link>
+                      </li>
+                    </>
+                  )
+                  : (
+                    <li className="header__nav-item">
+                      <Link
+                        className="header__nav-link"
+                        to={AppRoute.Login}
+                      >
+                        <span className="header__signout">Sign in</span>
+                      </Link>
+                    </li>
+                  )
+              }
+            </ul>
+          </nav>
+        </div>
+      </div>
+    </header>
+  );
+}
+
+export default Header;
