+++ src/pages/favorites-page/favorites-page.tsx
@@ -1,7 +1,7 @@
 import { Helmet } from 'react-helmet-async';
-import Logo from '../../components/logo/logo.tsx';
 import { Offer, OfferClick, OfferHover } from '../../types/offer.ts';
 import StayPlaceCards from '../../components/stay-place-card/stay-place-cards.tsx';
+import Header from '../../components/header/header.tsx';
 
 type FavoritesPageProps = {
   offers: Offer[];
@@ -16,32 +16,7 @@ function FavoritesPage({offers, onOfferClick, onOfferHover}: FavoritesPageProps)
       <Helmet>
         <title>6 cities. Favorites</title>
       </Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo />
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <Header />
 
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
