+++ src/pages/offer-page/offer-page.tsx
@@ -1,5 +1,4 @@
 import { Helmet } from 'react-helmet-async';
-import Logo from '../../components/logo/logo.tsx';
 import { Offer, OfferClick, OfferHover } from '../../types/offer.ts';
 import { useParams } from 'react-router-dom';
 import ReviewsList from '../../components/reviews/reviews-list.tsx';
@@ -10,6 +9,7 @@ import { Cities, OffersClassNames, STARS } from '../../const.ts';
 import { store } from '../../store/index.ts';
 import { useAppSelector } from '../../hooks/index.ts';
 import Loader from '../../components/loader/loader.tsx';
+import Header from '../../components/header/header.tsx';
 
 type OfferPageProps = {
   selectedOffer: Offer | undefined;
@@ -57,32 +57,7 @@ function OfferPage({selectedOffer, onOfferClick, onOfferHover}: OfferPageProps):
         <Helmet>
           <title>6 cities. Offer</title>
         </Helmet>
-        <header className="header">
-          <div className="container">
-            <div className="header__wrapper">
-              <div className="header__left">
-                <Logo />
-              </div>
-              <nav className="header__nav">
-                <ul className="header__nav-list">
-                  <li className="header__nav-item user">
-                    <a className="header__nav-link header__nav-link--profile" href="#">
-                      <div className="header__avatar-wrapper user__avatar-wrapper">
-                      </div>
-                      <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                      <span className="header__favorite-count">3</span>
-                    </a>
-                  </li>
-                  <li className="header__nav-item">
-                    <a className="header__nav-link" href="#">
-                      <span className="header__signout">Sign out</span>
-                    </a>
-                  </li>
-                </ul>
-              </nav>
-            </div>
-          </div>
-        </header>
+        <Header />
 
         <main className="page__main page__main--offer">
           <section className="offer">
